requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
executing==2.1.0
flask==3.0.3
frozenlist==1.4.1
gotrue==2.8.1
greenlet==3.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.9
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.6.2
jupyter-core==5.7.2
markupsafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
openai==1.45.0
orjson==3.10.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pip==24.0
platformdirs==4.3.3
postgrest==0.16.11
prompt-toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.9.1
pydantic-core==2.23.3
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyyaml==6.0.2
pyzmq==26.2.0
realtime==2.0.2
requests==2.32.3
ruff==0.6.5
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
storage3==0.7.7
strenum==0.4.15
supabase==2.7.4
supafunc==0.5.1
tenacity==8.5.0
tiktoken==0.7.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==12.0
werkzeug==3.0.4
yarl==1.11.1
markdown==3.7
python-slugify==8.0.1